Exit Slideshow to Main Menu Sprint 2
Able to exit slideshow via button or "ESC" key pressed. It will then
return to the Main Menu. Done in JavaFx whereby the individual GUI of
the Main Menu and Slide are done by Scene Builder where it's individual
layout are being stored in MainMenu.fxml and Slide.fxml respectively.