Main Menu JUnit Testing
Have fixed the mainMenTest class so that it doesn't contain any errors.
Have modified it to check that it is correctly producing a scene, group,
added some content to the group and setting the stage size to full
screen. Unfortunately I couldn't find a way to check the content of the
group more directly to test what had actually been added, for now it
just checks that it contains an HBox.