Set and reset single timer
Can now set a timer using a drop down list view. The timer can be
started paused and restarted using the same button. The reset buttons
stops the timer and returns it to the users original time setting. Upon
finishing the text changes to "Finished" and plays  sound to notify the
user it has finished.