Hideable Menu bar and Text Handler Timer Changes
Slide menu bar disappears when mouse cursor is not moving for 3 seconds.
Reappears after moving the mouse, stays visible when mouse cursor is
over the slide menu bar box. Added pause button to slide menu bar.
Changed timer threads in TextHandler to use Timelines to allow for
pausing and resuming the starttime and duration counts. When pressed
pauses all timelines in all instances of TextHandler, changes text to
resume. When pressed again, Timelines are restarted, currently unsure as
to if the Timelines are resuming correctly from where they were stopped.