Fixed bug: slideshow now returns to main menu past last slide or before first slide
Added logic to return to the main menu when the prev. slide button is
pressed (or the left key) and the first slide is showing.

Added logic to prompt the user to return to the main menu when the last
slide is showing and the next slide button is pressed (or the right
key).