Completed feature: add, remove and edit shopping list
Added new edit and “add” buttons to the shopping list GUI.
Edit allows the user to deselect any items from the list which will be
deleted when the save button is pressed.
Add allows the user to type items to add to the list.
Using the enter key speeds up adding items to the list.

Added validation to ShoppingList Class and method to add multiple items
in one call.